Corrections algorithme

+++ algorithme.h
@@ -16,9 +16,9 @@
 #include <arm_math.h>
 #include <params.h>
 
-float32_t HeartRate(float32_t *Signal, uint32_t temps1, uint32_t temps2);
-float32_t AC(float32_t *Signal, uint32_t temps1, uint32_t temps2);
-float32_t DC(float32_t *Signal, uint32_t temps1, uint32_t temps2);
-float32_t calculSpO2(float32_t *InputRed, float32_t *InputIR, uint32_t temps1, uint32_t temps2);
+float32_t HeartRate(uint32_t *Signal, uint32_t temps1, uint32_t temps2);
+float32_t AC(uint32_t *Signal, uint32_t temps1, uint32_t temps2);
+float32_t DC(uint32_t *Signal, uint32_t temps1, uint32_t temps2);
+float32_t calculSpO2(uint32_t *InputRed, uint32_t *InputIR, uint32_t temps1, uint32_t temps2);
 
 /* [] END OF FILE */
